+++ R/generics_coef.R
@@ -3,7 +3,7 @@
 #' @srrstats {G2.1a} Ensures that the input object is of the expected class (`apes`, `feglm`, or `felm`).
 #' @srrstats {G3.1a} Outputs coefficients in a consistent format for interpretability.
 #' @srrstats {G3.1b} Supports multiple model object types, maintaining a standardized interface.
-#' @srrstats {G3.1c} Provides access to summary statistics (`cm`) where applicable.
+#' @srrstats {G3.1c} Provides access to summary statistics (`coefficients`) where applicable.
 #' @srrstats {G5.1} Includes robust error handling for unsupported or invalid input objects.
 #' @srrstats {G5.4a} Includes tests for extracting coefficients from simple and complex model objects.
 #' @srrstats {RE4.2} Returns coefficients via a standard method for feglm-type objects and derived classes (i.e., felm, apes, etc).
@@ -33,17 +33,17 @@ coef.felm <- function(object, ...) {
 #' @export
 #' @noRd
 coef.summary.apes <- function(object, ...) {
-  object[["cm"]]
+  object[["coefficients"]]
 }
 
 #' @export
 #' @noRd
 coef.summary.feglm <- function(object, ...) {
-  object[["cm"]]
+  object[["coefficients"]]
 }
 
 #' @export
 #' @noRd
 coef.summary.felm <- function(object, ...) {
-  object[["cm"]]
+  object[["coefficients"]]
 }